gcs: Support uploading tmp chunks to separate tmp bucket
Currently we upload tmp chunks to the destination bucket and then remove
them after. This clashes with bucket-wide retention locks.

A bucket retention lock is a mitigation against ransomware attacks. It
disallows deletion of objects for a configurable interval.

By uploading tmp chunks to a separate tmp bucket, we can configure that
tmp bucket without retention lock. The composite objects are then copied
to the destination bucket, and deleted from the tmp bucket.

This incurs two additional GCS operations per uploaded object.

Note that it is not possible to compose objects across bucket
boundaries, this is why the copy operation is needed in addition to the
compose operation. The copy happens remotely.